YukyuData-AppFusion2.13/backend/requirements.txt
# =============================================================================
# YuKyuDATA-app - Dependencias Python
# =============================================================================
# Ultima actualizacion: 2026-01-10
# Python requerido: 3.10+
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi>=0.109.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.31.0
pydantic>=2.5.3,<3.0.0
starlette>=0.35.0,<0.42.0

# -----------------------------------------------------------------------------
# Excel Processing
# -----------------------------------------------------------------------------
openpyxl>=3.1.2,<4.0.0
XlsxWriter>=3.1.9,<4.0.0

# -----------------------------------------------------------------------------
# PDF Generation
# -----------------------------------------------------------------------------
reportlab>=4.0.0,<5.0.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
PyJWT>=2.8.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.1.2,<5.0.0  # Direct bcrypt for auth.py password hashing

# -----------------------------------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil>=2.8.2,<3.0.0
python-magic>=0.4.27,<0.5.0
psutil>=5.9.8,<6.0.0  # System monitoring for project status dashboard
cryptography>=42.0.0,<43.0.0

# -----------------------------------------------------------------------------
# Async Support
# -----------------------------------------------------------------------------
anyio>=4.0.0,<5.0.0
httpx>=0.26.0,<0.28.0

# =============================================================================
# Development & Testing
# =============================================================================

# -----------------------------------------------------------------------------
# Testing Framework
# -----------------------------------------------------------------------------
pytest>=7.4.4,<9.0.0
pytest-asyncio>=0.23.3,<0.25.0
pytest-cov>=4.1.0,<6.0.0

# -----------------------------------------------------------------------------
# Code Quality
# -----------------------------------------------------------------------------
# Uncomment for development
# black>=23.12.0,<25.0.0
# isort>=5.13.0,<6.0.0
# flake8>=7.0.0,<8.0.0
# mypy>=1.8.0,<2.0.0

# -----------------------------------------------------------------------------
# Security Scanning
# -----------------------------------------------------------------------------
# Uncomment for CI/CD
# bandit>=1.7.7,<2.0.0
# safety>=2.3.5,<4.0.0

# =============================================================================
# Optional Dependencies (uncomment if needed)
# =============================================================================

# -----------------------------------------------------------------------------
# Database Migrations & ORM (FASE 3)
# -----------------------------------------------------------------------------
alembic>=1.13.1,<2.0.0
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0  # PostgreSQL adapter for SQLAlchemy

# -----------------------------------------------------------------------------
# Caching
# -----------------------------------------------------------------------------
# redis>=5.0.1,<6.0.0

# -----------------------------------------------------------------------------
# Advanced Data Analysis
# -----------------------------------------------------------------------------
pandas>=2.1.4,<3.0.0
numpy>=1.26.0,<2.0.0

# -----------------------------------------------------------------------------
# Image Processing (for file validation)
# -----------------------------------------------------------------------------
# pillow>=10.2.0,<11.0.0

# -----------------------------------------------------------------------------
# E2E Testing
# -----------------------------------------------------------------------------
# playwright>=1.40.0,<2.0.0

# -----------------------------------------------------------------------------
# API Documentation
# -----------------------------------------------------------------------------
# (Included with FastAPI)

# -----------------------------------------------------------------------------
# Performance Monitoring
# -----------------------------------------------------------------------------
# prometheus-client>=0.19.0,<1.0.0
# opentelemetry-api>=1.22.0,<2.0.0
# opentelemetry-sdk>=1.22.0,<2.0.0

# =============================================================================
# Notes
# =============================================================================
#
# Para instalar todas las dependencias:
#   pip install -r requirements.txt
#
# Para instalar con dependencias de desarrollo:
#   pip install -r requirements.txt -r requirements-dev.txt
#
# Para actualizar dependencias:
#   pip install --upgrade -r requirements.txt
#
# Para verificar vulnerabilidades:
#   pip install safety && safety check
#
# =============================================================================
